+++ src/main/java/com/pythonstrup/demo/domain/user/repository/UserRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     // JPA Query Method
-    Optional<User> findById(int id);
+    Optional<User> findById(Long id);
     Optional<User> findByUsername(String username);
 }
